Move RedshiftController to controller.py
